src/test/javascript/spec/app/entities/cmt-type/cmt-type-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { Observable, of } from 'rxjs';

import { AssetMenagementCgesTestModule } from '../../../test.module';
import { CmtTypeUpdateComponent } from 'app/entities/cmt-type/cmt-type-update.component';
import { CmtTypeService } from 'app/entities/cmt-type/cmt-type.service';
import { CmtType } from 'app/shared/model/cmt-type.model';

describe('Component Tests', () => {
    describe('CmtType Management Update Component', () => {
        let comp: CmtTypeUpdateComponent;
        let fixture: ComponentFixture<CmtTypeUpdateComponent>;
        let service: CmtTypeService;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [AssetMenagementCgesTestModule],
                declarations: [CmtTypeUpdateComponent]
            })
                .overrideTemplate(CmtTypeUpdateComponent, '')
                .compileComponents();

            fixture = TestBed.createComponent(CmtTypeUpdateComponent);
            comp = fixture.componentInstance;
            service = fixture.debugElement.injector.get(CmtTypeService);
        });

        describe('save', () => {
            it(
                'Should call update service on save for existing entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new CmtType(123);
                    spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.cmtType = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.update).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );

            it(
                'Should call create service on save for new entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new CmtType();
                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.cmtType = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.create).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );
        });
    });
});
